:hammer:

+++ Projeto dieta/src/Validacoes/ValidaDados.java
@@ -2,7 +2,7 @@ package Validacoes;
 
 public class ValidaDados {
     
-    public boolean isAge (byte idade) {
+    public boolean isAge (int idade) {
         
         if(idade < 1 || idade > 120) {
             System.out.println("Idade inválida");
@@ -11,7 +11,7 @@ public class ValidaDados {
         return true;
     }
     
-    public boolean isHeight (short altura) {
+    public boolean isHeight (int altura) {
         if(altura < 35 || altura > 230){
             System.out.println("Altura inválida");
             return false;
@@ -19,7 +19,7 @@ public class ValidaDados {
         return true;
     }
     
-    public boolean isWeight (short peso) {
+    public boolean isWeight (int peso) {
         if(peso < 30 || peso > 240){
             return false;
         }
